Define a generic multistage channel selection filter.
My understanding is that this will be more efficient than a single
decimating low-pass filter for a given transition width.

I also get the impression that this can be implemented directly using
a single polyphase filterbank block, but I tried to write that and got
no useful output, whereas this appears to work.