Move OsmoSDR source to a "plugin" module.
This means that if one does not have gr-osmosdr installed, source.py
will not fail to import it and one can use other sources.

Note that despite the location it is not a plugin (in that it is not
loaded via twisted.plugin), since sources are hardcoded in the
configuration.