Add visible client audio state and revise buffer-sizing algorithm.
The audio buffer is now measured in samples instead of chunks
(which were arbitrary) and the target buffer size is determined
continuously and directly from the min/max buffer fullness. The old
algorithm seemed to be wrong; e.g. it would only _decrease_ the queue
limit in the event of an _overrun_.

All the console spam about overrun/underrun is replaced with the info
object. In the future there might be settings such as mute, volume, or
preferences about latency.