Use separate spectrum_fft msg_queue for every state stream.
Previously, in the presence of multiple state streams each FFT frame
would go only to whichever one of them grabbed it first; now all of
them get it.

Note that StreamCell.get now returns uninterpreted bytes rather than any
more sensible value -- but we never actually use that value, so it's not
worth fixing for now.

In a future revision I hope to arrange so that the FFT processing itself
is shut off when no client is actually looking at it, and the frame rate
(and possibly averaging) is selected based on client request (much like
the audio stream); this does none of that but is a step in the right
direction.