Disable broken spin buttons on Knob widget.
I tested the revision in which I originally added them and they don't
work. Therefore the simplest explanation is that something changed in
Chrome's click/focus handling which interacts poorly with the button
auto-hide strategy.

For now, I'm leaving the code in as a reminder that we used to have
this functionality and a guide to reimplementing it.