Add frequency transform parameters to OsmoSDRSource constructor.
This is a workaround for the the current inability to dynamically change
an exported field's type (the frequency range), allowing them to be
initialized early enough, in the configuration file, to take effect.
Note that the range can still get out of sync if the setting is
changed in the UI (since that will override whatever goes into the
constructor); this is purely a kludge to make using the frequency
shift feasible.