Overlapped FFT support for greater time resolution.
Due to gnuradio.blocks.stream_to_vector_decimator, used by logpwrfft,
not natively supporting overlapped output, we use a pre-processing stage
which duplicates samples going in to it.

This is an inefficient solution, especially as it means the overlap
cannot be changed without modifying the flowgraph (thus we set it to
a worst-case value, copying data that is then discarded when at low FFT
rates). It would be better to have a stream_to_vector_decimator block
which internally supported overlapping; the only reason I have not done
so is to avoid including C++ code (and thus having a mandatory build
step).