Move spectrum details to a new block 'MonitorSink'.
Also, in the UI the rate/resolution controls are now displayed in the
spectrum header rather than buried under Top's "more".

This change is primarily intended to make it more feasible to drop in
additional spectrum displays, such as for audio output or internal
stages of a demodulator. Further work needed is to have a corresponding
client-side widget (what's currently hardcoded in index.html) and
perhaps allowing the server to not send FFT frames for monitors the
client isn't actually displaying in the UI.