Move cell-polling from StateStreamInner to a separate class values.Poller.
This is refactoring to support a number of future features, none of
which are implemented here:
* support non-polling cell update notifications
* adjustable polling for rarely-changing values
* multiple clients don't need to poll independently
* clients can express interest or lack thereof on each object

The Poller object provides an interface intended to work as if the
individual cells had actual change notifications. StateStreamInner is
now defined in terms of getting messages from the Poller rather than
always traversing every reachable object every polling time.

Also added some tests; particularly of what happens if the same object
is reachable by multiple paths (since we're not traversing everything
any more, we have to use refcounts (of the on-the-wire references)
rather than deleting every object we didn't see this time).

The revised code is a bit sloppy with names for the internals; I expect
to be changing it anyway for the above features.