Fix handling of reconnecting to server.
The so-called "radio" (top block) object was saved from the first
connection and reused, leading to various things getting out of
sync or not working at all. Now we always retrieve it from a cell.

For the excessively curious, this has been broken since Oct 2013,
commit 4128645297a6c2e0b96e765658d7a8e9243feac3, when network.js
was changed so that reconnecting produced a new root block object
rather than reshaping the existing one.