Add a warning message when the source signal amplitude appears to be too high.
This is intended to avoid the user needing to know beforehand that
clipping can occur, or to recognize that it is occurring from the
spectrum plots. It does have the built-in assumption that the maximum
sample magnitude is 1.0, which is not true e.g. for SimulatedDevice,
but as far as I know is true for all GNU Radio hardware drivers.