Add squelch options to APRS demodulator, defaulting to Voice Alert.
Voice Alert is a protocol for APRS use where attended stations transmit a
CTCSS tone along with their packets, and use it for initiating simplex
voice contacts. <http://www.aprs.org/VoiceAlert3.html>

It is currently fixed to recognizing only the USA 100 Hz tone.

In order to detect the CTCSS tone, I had to add a hook to remove the
bandpass filter from FMDemodulator (which was added to remove CTCSS
tones from audio output). This does not seem the best way to go about
it, but was simple. A better solution might be to add a "discriminator
tap" feature, i.e. a second output which is less filtered and, once the
regular FM demod has CTCSS squelch, not squelched.