Draw WaterfallPlot on window-sized canvas instead of FFT-sized canvas.
This requires the 2D-canvas waterfall to add an extra resizing step, but
it is a step towards more flexibility, in particular being able to add a
zoomed-in FFT or simply support for more than 4096 points (which is a
common restriction on the maximum dimensions of canvases).

Bonus feature: 2D-canvas waterfall does not clear on resolution change.

Negative effects:
* 2D-canvas waterfall is now significantly slower to do a full repaint
  (e.g. while zooming).
* WebGL waterfall requires OES_float_linear extension to use float
  textures (because the history texture is non-power-of-two).

Supporting changes:
* Since WaterfallPlot and ScopePlot scale the canvas the same way, the
  code for that is now in CanvasSpectrumWidget.
* boundedFn now forwards arguments (but not return value).