Reconfigure SSB AGC for reasonable output levels.
Previously, the AGC was nearly always maxed out and useless. This was
based on a half-baked idea: that the gain + AGC configuration could be
set so that 'no signal' did not mean listening to the amplified noise
floor, and the primary function of the AGC was to protect your ears from
sudden strong signals. This doesn't really work unless we were to add
some sort of detector for the noise floor level, since it varies across
the spectrum -- the results we got were mainly that SSB was always too
quiet and so switching to AM or FM blasts your ears.

One possible way to recover the 'ear protector' idea would be to add a
fast AGC after a slow AGC, such that the fast AGC is usually maxed out
(having a larger reference level than the slow AGC), but I have tested
the current configuration and find it adequate.