Change automatic tuning to be more centered.
The original design was left over from being used in a scanning function
that no longer exists, intended to minimize the number of center
frequency changes. Centering will give better signal in almost all
cases, and is less broken in wideband cases like Mode S.

We're still assuming a DC offset that might not be present.