OsmoSDRProfile: add and use information on nonfunctional features.
For many gr-osmosdr devices, set_dc_offset_mode, set_iq_balance_mode,
and set_gain_mode don't actually do anything, and there's no interface
to find this out other than seeing if the setter returns the new value
(and some fake that). So, add flags for this to OsmoSDRProfile.

The primary benefit is that devices which do not support AGC will not
show an “AGC On” button that does nothing. The dc_cancel checkbox is
also subject to this. iq_balance ought to be too, but in practice that
means figuring out if gr-osmosdr was compiled against gr-iqbalance,
so I'm putting that off.

To express this information to the client, added a “Constant” ValueType
(so _OsmoSDRRXDriver can choose Constant(False) instead of bool).

Also added to the profile a “supports TX” flag, for later.