Add entire-input-band mode to AMDemodulator.
This mode is primarily useful for airband listening: it allows
simultaneously listening to all AM transmissions in the hardware
passband, and has low noise compared to narrow-band AM without
squelch.

Ideally, this would be a specialized receiver which doesn't have any
specific channel frequency and hence is never invalid, but that
needs more design work.