Add Coordinator; move radio.tune (but not yet radio.preset) to it.
The Coordinator is planned to provide a hub for interactions between
parts of the UI, in particular those which have different “axes”
(the map, the spectrum view, and the widget tree), in such a way that
we do not need everything to know about everything else (and main.js
not to need so much hardcoding).

For now, what it's doing is giving a place to put the tune() function
which is not a patch to the radio (a.k.a Top) object.