Reduce horizontal antialiasing samples in spectrum display from 17 to 3.
For some bizarre graphics-programming reason, the previous version
caused the fill-intensity to always be 0 as well as resulting in slow
performance, on a MacBook Pro AMD Radeon HD 6770M and Chrome 45.

The reduced sampling does make the lines have more gaps, but this is
not so bad now that we have the fill as a distraction.