Add popup menus to UI.
The menus are based on <dialog> for useful default behavior. (<menu> is
not currently supported in Chrome even if it would be suitable.)

Menu contents are widgets. Widgets can signal via Context to close an
enclosing menu if there is one.

The first use is that clicking on a grouped FreqScale entry will display
a menu with the individual entries.