Split out unselective AM demodulator and add stereo positioning.
The “use_entire_input_band” option on the AM demodulator is now
considered a separate mode and has a separate demodulator class.
Its output is stereo, and panned according to the RF frequency
(relative to the device center frequency) of the heard signal.
This works by using two demod chains each applying the new
“sawtooth” filter but with opposite slopes.