Show/hide subwindows using a sidebar menu rather than clickable headers.
This is in order to eliminate two issues with the previous design: the
show-buttons took up too much space in the header when many subwindows
were hidden, and on a touchscreen it was too easy to hit the wrong
target.

The new design is a “hamburger menu”, which, while not the best UI
paradigm around, fits nicely with the existing structure of the page
(being a vertical column on the left that can be hidden to take up no
extra space) and provides a better place to put things like the link
to the manual.

Future work might be to make the sidebar overlay instead of pushing
aside the main content when viewport width is scarce, and autohide.