Add database info for WWV, WWVH, and WWVB.
This information was obtained from the National Institute of Standards
and Technology's web site at <http://www.nist.gov/pml/div688/grp40/>,
“Time and Frequency Services”.

These broadcasts are receivable US-wide so they are not too local to
include in the builtin databases which already include US-specific
information, and they are our first examples of latitude/longitude
information in databases.

This is not a great fit for the "broadcast bands" file, but I don't want
to have a whole separate file for the purpose (yet).