Implement command cells.
A "command" is (currently) a zero-argument method that can be called
by the client. A command cell is a cell that serves to represent the
existence of the command (and just has a value of null).

This isn't all that clean a design in principle, but it has the
advantage of not requiring an entire mechanism separate from 'state'
to implement commands.