Add config.features, replacing config.set_stereo.
This interface allows us to add more feature controls without
proliferation of different methods. I intend to use it primarily to
allow work-in-progress features to be added to the master branch,
simplifying development in some cases and allowing me to alpha-test
multiple features at once.

config.set_stereo has been replaced with the 'stereo' feature but is
still available as an alias.