Tweak spectrum widget horizontal antialiasing again.
I believe that the _correct_ solution here is to use an intermediate
buffer which is the width of the widget instead of the width of the FFT,
which we can paint using a more correct antialiasing strategy and then
use as our input to the actual drawing stage without excessive
per-drawn-pixel costs. But for now, these parameters seem to be less
wrong, eliminating the visible double peaks in favor of blockier lines.