Change format for map feature line rendering.
The previous format assumed that every feature's line ended at that
feature's specified position, which is great for object tracking but not
so useful for everything else. The new format allows specifying an
arbitrary set of polylines (MultiLineString in GeoJSON terminology).

The direction I see this going is for the renderer function to return
an arbitrary list of lines and text-or-icon billboards and polygons and
whatnot, rather than a hardcoded "one icon, one line of text, and lines"
structure. This will allow support for, for example, APRS area symbols.