Instead of config files, have config directories.
(Config files are still supported for now.)

The advantage of a config directory is that there is a natural place for
auxiliary files such as databases and state.json -- and particularly,
for new files which were not yet imagined when the user created their
config.

Future work may include adding a config method for "everything which is
not read-only shall be written to this other directory", for users who
want to keep a configuration/state distinction.