Replace EnumRow "short_desc" and "long_desc" with "label" and "description".
Incidentally fixes a bug in the frequency record editor mode menu,
which was looking for .sdesc instead of .short_desc.

This terminology change is more consistent with the terminology I
already use elsewhere (frequency database records have a 'label')
and wish to use in future work. A 'label', in this sense, is a string
that some object has to identify it to people, but which does not
actually have any guarantee of uniqueness or other properties.

It also avoids suggesting that the 'long_desc' is a substitute for the
'short_desc' (whereas in reality the former should only be displayed
together with the latter).