Move some modules into an "implementation" package shinysdr.i.
While ShinySDR is not in general stable enough to have a "stable API"
at all yet, I feel like it is worthwhile to try to create a distinction
between interface and implementation, where by interface I mean anything
a plugin might use. In particular, I intend to feel more free to
reorganize things in the implementation section, and to create modules
that represent useful but not necessarily "meaningfully named" groupings
of code.

Moving these files requires updating the following sorts of references:

* import statements, of course
* MANIFEST.in
* package names occurring in string literals
* __file__-relative paths referring to deps/