Refactor and fix issues with web static content.
Some of the File objects were configured with our desired parameters,
some were not; now they all are.

The contents of / are now all explicitly created, rather than taking
the contents of shinysdr/i/webstatic/. This is for the sake of future
multisession support, which will be dynamically creating entries in /,
and generally so that the namespace usage is better-documented.