Add pause control to AudioAnalyzerAdapter.
It is always paused on load. This is a kludge to avoid wasting CPU/GPU
time since we have no easy way to make the analyzer and MonitorSink stop
when not visible and we don't intend this to be normally always-visible
like the RF spectrum view.

Also, pause when the audio stream is stopped, regardless of checkbox.