Replace Cell.isBlock with new Reference type.
This is refactoring towards eliminating the parallel structure of
block cells and value cells.

('Reference' is the equivalent of the client-side type 'block'; the name
difference is because 'block' is not really good terminology.
Unfortunately, 'reference' isn't great either outside of the specific
context of 'this cell value is by-reference across the network'.)