Create cell 'metadata', put 'type' and 'persists' inside it.
I plan to add more fields to the metadata which will be used to
configure the client-side view of the cells (e.g. 'human readable'
label text). Keeping all these things in one pure-data structure
will keep things simple and orthogonal, for the most part.

The caveat is that the cell constructor parameters are remaining
separate fields. This is because we want to be able to write
    @exported_state(type=Range(...), label='Gain', ...)
rather than having to write out the metadata structure specifically,
and because some metadata fields may have default values chosen based
on other parameters (as 'persists' does now).

I have also taken this opportunity to bring cell description()s
in line with the vague JSON structure convention used elsewhere:
the field previously called 'kind' is now called 'type' and its
values mention that they are a cell. The old 'type' field is now
metadata.value_type.