For stereo audio, use vectors instead of independent ports.
Each hier block that produces or consumes stereo audio now represents
it as one port with vector-of-2-floats items rather than two separate
ports.

The old representation, while more conventional, has the disadvantage
that when the flow graph is reconfigured, if the (multithreaded!)
processing of the two channels happens not to be at exactly the same
position, a delay between the two channels may be introduced. By
explicitly pairing up the left and right elements, except within
subgraphs which are either not modified or destroyed and recreated in
their entirety, this skew is prevented.

Incidentally, the resampler cache in shinysdr.i.audiomux.BusPlumber
has been removed. If we want to resurrect its functionality, we should
do so by caching the filter tap calculations rather than the blocks,
which is more broadly applicable and does not reuse old possibly-skewed
output buffers.