Fix/optimize AGC UI.
The bug was that a transition from agc=false to agc=true not done by
the same client would be reset by the client.

(The setIfDifferent is not part of the fix, just to optimize some
gratuitous notification triggering; the bug was in the condition
    mode === 'auto' && !block.agc.get()
which if false meant either "going to manual gain mode" or "do not
need to set AGC cell" but the else-if assumed that it meant the
former.