Replace boundedFn's job with SubScheduler.
The idea of boundedFn is that widgets explicitly use it to break their
`function f() { cell.depend(f); }` notification loops when the widget is
destroyed. But I never consistently used it (partly due to thinking that
I had already implemented and deployed SubScheduler), so this should
be more robust, because SubScheduler has the same effect on all uses of
`config.scheduler` inside a widget.

There was also a bug in the destruction code: the lifecycledestroy event
listener was capturing, meaning that it responded to child widgets being
destroyed. This didn't cause anything to visibly break only because
boundedFn was used so little.