Refactor SpectrumView to allow other uses.
The more general idea is an object which gives widgets information
about their parent elements so as to know how to render themselves,
but not necessarily in the RF spectrum display. The generalization
is called "layout context".

Planned uses:
* Tabular layout of widgets: the layout context tells the widget what
  column(s) to show.
* Possibly convert the map to use HTML labels.

Decoupling the specifics of SpectrumView also allows moving
the code into spectrum.js which makes much more sense.