Replace audio disconnection with suspend()/resume().
This is not only a more direct way to express that we just want to stop
spending resources on playing silence, but also aligns better with
Chrome's upcoming autoplay-blocking policy:
https://developers.google.com/web/updates/2017/09/autoplay-policy-changes#webaudio

What we do not yet have for that policy is a way to resume() after an
interaction if the stream is already active on page load.

This commit also removes the 'lockout' feature from AudioAnalyserAdapter
because it is no longer used.