Move flake8 config entirely to comand line options.
For some reason with the flake8 I currently have installed the config
is ignored. Making this change so I can make progress, and because
apparently things like this have been an issue in the past (see for
example <https://stackoverflow.com/q/28436382> from 2016), and because
the config was already split across files.

The advantage of using setup.cfg would be that manually running flake8
alone will pick up the project configuration.