Remove SquelchMixin from SSB demodulator.
The removal is accomplished by removing SquelchMixin from
SimpleAudioDemodulator's superclasses and adding it back to the
individual demodulators.

The problem is that squelch followed by slow AGC means that the AGC
tries to compensate for the squelch resulting in loud bursts when
the squelch opens. This could be fixed by routing the squelch muting
around the AGC or the AGC gain around the squelch, but this can't
be done without putting together several primitive blocks instead
of using the stock squelch or AGC, which will decrease performance.

An alternative would be to remove AGC instead of removing squelch,
but this is not clearly an improvement as it would require users to
adjust gain manually. I am considering adding options for a variety
of processing modes including "AGC on the pre-channel-filter signal",
but that would want most of this change anyway.