Use WEBGL_lose_context to ensure cleanup.
This change prevents seeing "too many WebGL contexts" warnings in the
console upon repeatedly toggling the "Enable OpenGL" setting.

Arguably that we have to do this is a bug because it points to lack of
garbage collection of widgets (almost certainly due to miscellaneous
Notifiers holding onto callbacks that haven't fired) but at least that's
not holding on to GPU resources rather than just memory.